Add SnackBar Extension to Export Manager
This commit integrates the `SnackBarExtension` into the `export_manager.dart` file, allowing for easier access to the extension across the project. The `SnackBarExtension` is now part of the `export_manager.dart` file, which simplifies the import process and ensures consistency in how extensions are managed.

The `SnackBarExtension` provides a convenient method to display snack bars within the context of a `BuildContext`, enhancing the usability of the extension throughout the application.