Add language type extension and translation helper
This commit introduces a new file, `language_type_extension.dart`, which contains extensions for handling language types and translations. The `LanguageTypeExtension` extension provides a method to get the corresponding language code for `LanguageType` enums. Additionally, the `TranslateX` extension on `String` allows for easy translation of strings using the `AppLocalizations` class.

The constants defined in the file include language codes for Arabic and English, as well as their respective locales. The `assetPathLocal` constant specifies the path to translation assets.