Add GlobalKey for NavigatorState to MyApp Widget
This commit introduces a `GlobalKey<NavigatorState>` to the `MyApp` widget, which is used to manage the navigation state of the application. The `navigatorKey` is then passed to the `MaterialApp` widget, allowing for more controlled navigation throughout the app. This change is necessary for scenarios where global navigation control is required, such as programmatically pushing routes from outside the widget tree.